Integrate esbuild to support old node versions
Will look into integrate this with TypeScript at some point afterwards.
